requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
spacy==3.7.2
requests==2.31.0
gTTS==2.4.0
moviepy==1.0.3
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
uuid==1.30
pydub==0.25.1 